+++ lib/slackify/handlers.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'handlers/base'
-require_relative 'handlers/configuration'
+require_relative 'handlers/factory'
 require_relative 'handlers/unhandled_handler'
 require_relative 'handlers/validator'