+++ lib/slackify/handlers/unhandled_handler.rb
@@ -2,6 +2,10 @@
 
 module Slackify
   module Handlers
+    # Default handler for any text message that is not handler by other handlers
+    # This can easily be replaced by setting you own unhandled handler when
+    # configuring slackify. Use `#unhandled_handler=` to set it. Your handler
+    # must implement a `self.unhandled` method.
     class UnhandledHandler < Base
       def self.unhandled(params)
         slack_client.chat_postMessage(