+++ test/dummy/config/initializers/01_handlers.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-Slackify.load_handlers